requirements.txt
-i https://pypi.org/simple
alembic==1.15.2; python_version >= '3.9'
bcrypt==4.3.0; python_version >= '3.8'
bidict==0.23.1; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
click==8.1.8; python_version >= '3.7'
flask==3.1.0; python_version >= '3.9'
flask-bcrypt==1.0.1
flask-cors==5.0.1; python_version >= '3.9' and python_version < '4.0'
flask-migrate==4.1.0; python_version >= '3.6'
flask-socketio==5.5.1; python_version >= '3.6'
flask-sqlalchemy==3.0.5; python_version >= '3.7'
greenlet==3.2.1; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.16.0; python_version >= '3.8'
importlib-metadata==6.0.0; python_version >= '3.7'
importlib-resources==5.10.0; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
mako==1.3.10; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
python-dotenv==1.1.0; python_version >= '3.9'
python-engineio==4.12.0; python_version >= '3.6'
python-socketio==5.13.0; python_version >= '3.8'
simple-websocket==1.1.0; python_version >= '3.6'
sqlalchemy==1.4.49; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy-serializer==1.4.1
typing-extensions==4.13.2; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'
wsproto==1.2.0; python_full_version >= '3.7.0'
zipp==3.21.0; python_version >= '3.9'
